+++ webservice/src/api/mod.rs
@@ -29,9 +29,9 @@ async fn check_corpora_authorized_read(
         return Ok(requested_corpora);
     }
 
-    let conn = db_pool.get()?;
+    let mut conn = db_pool.get()?;
     let allowed_corpora =
-        web::block(move || actions::authorized_corpora_from_groups(&claims, &conn)).await??;
+        web::block(move || actions::authorized_corpora_from_groups(&claims, &mut conn)).await??;
     if requested_corpora
         .iter()
         .all(|c| allowed_corpora.contains(c))